refactor: change location phone code country interface
