+++ airways/src/app/services/phone-code-name.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { PhoneCodeCountry } from '../constants/phone-code-country';
-import { PhoneCodeCountryInterface } from '../interfaces/phone-code-country.interface';
+import { PhoneCodeCountryInterface } from '../shared/models/phone-code-country.interface';
 
 @Injectable({
   providedIn: 'root',
